Move args parsing out of stats.py
