+++ ghcl/github_stats.py
@@ -3,7 +3,7 @@ from typing import List
 
 import requests
 
-from client.models.pull_request import PullRequest
+from ghcl.models.pull_request import PullRequest
 
 
 class GithubStats:
